# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FarmFlow API
  description: |
    Egyptian livestock marketplace API.
    All monetary values are in EGP (Egyptian Pound).
    All dates are ISO 8601 strings.
    Image paths are relative to the server root (e.g. `/uploads/filename.jpg`).

    **Authentication:** Pass a JWT token in the `Authorization: Bearer <token>` header.

    **Roles:**
    - `buyer` — can browse, order, review, manage favorites
    - `seller` — can manage listings, animals, dairy, supplies, finances
    - `admin` — full access including approval workflows

    **⚠️ Mobile app endpoint notes:**
    - `GET /api/listings/my` is a dedicated route (seller/admin JWT required) — returns own listings (all statuses)
    - `GET /api/orders/my` is handled by `GET /api/orders` (buyer JWT → own orders only)
    - `PATCH /api/listings/:id/status` does not exist — use `PUT /api/listings/:id` with a `status` body field
    - `PATCH /api/admin/users/:id` does not exist — use `PATCH /api/admin/users/:id/toggle`
  version: 1.0.0
  contact:
    email: a.khalil@soar.inc
servers:
  - url: http://localhost:5001/api
    description: Local development
  - url: https://api.farmflow.app/api
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MongoId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
      example: 64f1a2b3c4d5e6f7a8b9c0d1
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Unauthorized
...
    Animal:
      type: object
      properties:
        _id:    { $ref: '#/components/schemas/MongoId' }
        seller: { $ref: '#/components/schemas/MongoId' }
        tagId:  { type: string }
        type:   { type: string, enum: [cattle, buffalo, sheep, goat, camel, horse, poultry, rabbit, other] }
        breed:  { type: string }
        gender: { type: string, enum: [male, female, unknown] }
        dob:    { type: string, format: date }
        color:  { type: string }
        currentWeight: { type: number }
        targetWeight:  { type: number }
        nextWeighingDate: { type: string, format: date }
        healthStatus: { type: string, enum: [healthy, sick, quarantine, deceased] }
        status:       { type: string, enum: [active, sold, deceased] }
        notes:        { type: string }
        images:       { type: array, items: { type: string } }
        pregnancyStatus: { type: string, enum: [none, pregnant, recently_gave_birth] }
        pregnancyDate:   { type: string, format: date }
        expectedBirthDate: { type: string, format: date }
        birthCount:      { type: integer }
        weightLog:
          type: array
          items:
            type: object
            properties:
              _id:      { $ref: '#/components/schemas/MongoId' }
              weightKg: { type: number }
              date:     { type: string, format: date-time }
              notes:    { type: string }
        vaccinationLog:
          type: array
          items:
            type: object
            properties:
              _id:         { $ref: '#/components/schemas/MongoId' }
              vaccine:     { type: string }
              date:        { type: string, format: date-time }
              nextDueDate: { type: string, format: date-time }
              vet:         { type: string }
              notes:       { type: string }
        createdAt: { type: string, format: date-time }
    AnimalSummary:
      type: object
      properties:
        total:        { type: integer }
        byType:       { type: object, additionalProperties: { type: integer } }
        avgAgeMonths: { type: number, nullable: true }
        avgWeightKg:  { type: number, nullable: true }
    WeightEntry:
      type: object
      properties:
        _id:      { $ref: '#/components/schemas/MongoId' }
        weightKg: { type: number }
        date:     { type: string, format: date-time }
        notes:    { type: string }
    VaccinationEntry:
      type: object
      properties:
        _id:         { $ref: '#/components/schemas/MongoId' }
        vaccine:     { type: string }
        date:        { type: string, format: date-time }
        nextDueDate: { type: string, format: date-time }
        vet:         { type: string }
        notes:       { type: string }
    MedicalRecord:
      type: object
      properties:
        _id:          { $ref: '#/components/schemas/MongoId' }
        animal:       { $ref: '#/components/schemas/MongoId' }
        seller:       { $ref: '#/components/schemas/MongoId' }
        date:         { type: string, format: date-time }
        diagnosis:    { type: string }
        treatment:    { type: string }
        medication:   { type: string }
        vet:          { type: string }
        cost:         { type: number }
        followUpDate: { type: string, format: date }
        notes:        { type: string }
        resolved:     { type: boolean }
        createdAt:    { type: string, format: date-time }

    # ── Finance ─────────────────────────────────────────────────────────────────
...
paths:
...
  /animals:
    get:
      tags: [Animals]
      summary: Get animals (seller sees own; admin sees all)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: type
          schema: { type: string }
          description: Filter by animal type
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
          description: Page number for server-side pagination (requires `limit`). When omitted, returns full array (backwards-compatible).
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
          description: Items per page. When `page` is also provided, enables paginated response format.
      responses:
        '200':
          description: Array of animals (plain array when no `page` param; paginated object when `page` is present)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: '#/components/schemas/Animal' }
                  - type: object
                    properties:
                      items:   { type: array, items: { $ref: '#/components/schemas/Animal' } }
                      total:   { type: integer }
                      page:    { type: integer }
                      pages:   { type: integer }
                      hasMore: { type: boolean }

    post:
      tags: [Animals]
      summary: Add animal to herd (seller or admin)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [type]
              properties:
                type:          { type: string, enum: [cattle, buffalo, sheep, goat, camel, horse, poultry, rabbit, other] }
                breed:         { type: string }
                tagId:         { type: string, description: Unique tag/ear number per seller }
                gender:        { type: string, enum: [male, female, unknown] }
                dob:           { type: string, format: date }
                color:         { type: string }
                currentWeight: { type: number }
                healthStatus:  { type: string, enum: [healthy, sick, quarantine, deceased] }
                notes:         { type: string }
                images:        { type: array, items: { type: string, format: binary } }
      responses:
        '201':
          description: Animal created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Animal' }
        '409':
          description: Tag ID already used for another animal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /animals/summary:
    get:
      tags: [Animals]
      summary: Get herd statistics summary
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Herd summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnimalSummary' }
  /animals/follow-ups-due:
    get:
      tags: [Animals]
      summary: Get medical records with follow-up date due within 7 days
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Array of medical records (with animal populated)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MedicalRecord' }
  /animals/weighing-due:
    get:
      tags: [Animals]
      summary: Get active animals with weighing date due within 7 days
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Array of animals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Animal' }
  /animals/{id}:
    get:
      tags: [Animals]
      summary: Get animal by ID
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      responses:
        '200':
          description: Animal detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Animal' }
        '403': { description: Forbidden }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

    put:
      tags: [Animals]
      summary: Update animal (seller owns it or admin)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                type:            { type: string, enum: [cattle, buffalo, sheep, goat, camel, horse, poultry, rabbit, other] }
                breed:           { type: string }
                gender:          { type: string, enum: [male, female, unknown] }
                color:           { type: string }
                currentWeight:   { type: number }
                targetWeight:    { type: number }
                nextWeighingDate: { type: string, format: date }
                healthStatus:    { type: string, enum: [healthy, sick, quarantine, deceased] }
                status:          { type: string, enum: [active, sold, deceased] }
                notes:           { type: string }
                pregnancyStatus: { type: string, enum: [none, pregnant, recently_gave_birth] }
                pregnancyDate:   { type: string, format: date }
                expectedBirthDate: { type: string, format: date }
                birthCount:      { type: integer }
                keepImages:      { type: string, description: JSON array of existing paths to keep }
                images:          { type: array, items: { type: string, format: binary } }
      responses:
        '200':
          description: Updated animal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Animal' }
        '403': { description: Forbidden }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

    delete:
      tags: [Animals]
      summary: Delete animal
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      responses:
        '204': { description: Deleted }
        '403': { description: Forbidden }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /animals/{id}/weight:
    post:
      tags: [Animals]
      summary: Add weight entry (updates currentWeight automatically)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [weightKg]
              properties:
                weightKg: { type: number, minimum: 0.1 }
                date:     { type: string, format: date-time }
                notes:    { type: string }
      responses:
        '201':
          description: Weight entry created; milestone notification sent if target reached
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WeightEntry' }
  /animals/{id}/weight/{entryId}:
    delete:
      tags: [Animals]
      summary: Delete weight entry
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
        - in: path
          name: entryId
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      responses:
        '204': { description: Deleted }
  /animals/{id}/vaccination:
    post:
      tags: [Animals]
      summary: Add vaccination record
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vaccine]
              properties:
                vaccine:     { type: string }
                date:        { type: string, format: date-time }
                nextDueDate: { type: string, format: date-time }
                vet:         { type: string }
                notes:       { type: string }
      responses:
        '201':
          description: Vaccination entry created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VaccinationEntry' }
  /animals/{id}/vaccination/{entryId}:
    delete:
      tags: [Animals]
      summary: Delete vaccination record
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
        - in: path
          name: entryId
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      responses:
        '204': { description: Deleted }
  /animals/{id}/medical:
    post:
      tags: [Animals]
      summary: Add medical record (auto-creates doctor expense when cost > 0)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:         { type: string, format: date-time }
                diagnosis:    { type: string }
                treatment:    { type: string }
                medication:   { type: string }
                vet:          { type: string }
                cost:         { type: number, minimum: 0 }
                followUpDate: { type: string, format: date }
                notes:        { type: string }
                resolved:     { type: boolean }
      responses:
        '201':
          description: Medical record created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MedicalRecord' }

    get:
      tags: [Animals]
      summary: List medical records for an animal
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      responses:
        '200':
          description: Array of medical records (sorted by date desc)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MedicalRecord' }
  /animals/{id}/medical/{recordId}:
    patch:
      tags: [Animals]
      summary: Update medical record (e.g. mark resolved)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
        - in: path
          name: recordId
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                resolved:     { type: boolean }
                followUpDate: { type: string, format: date }
                notes:        { type: string }
      responses:
        '200':
          description: Updated record
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MedicalRecord' }
        '404':
          description: Record not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

    delete:
      tags: [Animals]
      summary: Delete medical record
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
        - in: path
          name: recordId
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      responses:
        '204': { description: Deleted }

  # ── FINANCE ──────────────────────────────────────────────────────────────────
...